tune failure rate based on parallelism

+++ flink-benchmarks/src/main/java/flink/benchmark/BenchmarkConfig.java
@@ -63,7 +63,7 @@ public class BenchmarkConfig implements Serializable {
     public final String singlelevelPath;
     public final String singlelevelStateBackend;
     public final int maxMemStateSize;
-    public final double mttiMs;
+    public final long mttiMs;
 
     /**
      * Create a config starting with an instance of ParameterTool
@@ -126,7 +126,7 @@ public class BenchmarkConfig implements Serializable {
 
         this.maxMemStateSize = parameterTool.getInt("max.memory.state.size", 5242880);
 
-        this.mttiMs = parameterTool.getDouble("mtti.ms", 20_000);
+        this.mttiMs = parameterTool.getLong("mtti.ms", 20_000);
 
     }
